changed code to handle the "connection already in progress" error and optimized handling of non-blocking sockets
select() method removed for a simpler use of non blocking sockets in the serverPC part.
adapted error handling due to the removal of this method.
Changed all close() method for closesocket(), the later closing properly the socket when quitting.
changed and adapted printfs for more understanding and clarity when using the app.
Added a function to print warnings in the 3DS.
The function ReadDSInputInfo() now return -1 when  the error flags not covered by the code are sets, exiting the apps.
accept_new_connection() function removed